Refactored component name
