+++ src/app/home-bar/home-bar.module.ts
@@ -5,7 +5,7 @@ import {BrowserModule} from '@angular/platform-browser';
 import {NgModule} from '@angular/core';
 import {RouterModule} from '@angular/router';
 import {HttpModule} from '@angular/http';
-import {MODULE_COMPONENTS, MODULE_ROUTES} from './button-bar.routes';
+import {MODULE_COMPONENTS, MODULE_ROUTES} from './home-bar.routes';
 
 @NgModule({
   imports: [BrowserModule, HttpModule, RouterModule.forChild(MODULE_ROUTES)],
