+++ src/app/home-bar/index.ts
@@ -1,4 +1,4 @@
 /**
  * Created by alexponce on 4/26/17.
  */
-export * from './button-bar.module';
+export * from './home-bar.module';
